gameWrapper component built out
